(Partial) 1.17 Update
Most of the needed updates are done, the only part that is missing is ChorusLinkUtils#doChorusLinkSearch, the list blockEntities list got removed and I decided to not touch it and discuss a viable good method to find the nearest chorus link first.
An ok idea would be to gradually search in the nearest chunks (because the blockEntities list is still present there) to the player until the max range of block is covered, compile a list of the various links found and find the closest one.